embedding-service/requirements.txt
fastapi
uvicorn[standard]
sentence-transformers
pydantic
numpy
python-dotenv